Circle fixes: extract "set error" thunks
